ci: only compile web container once
As the web-version is cross compiled to wasm anyway,
there is no need to build the container on arm again.
Instead, the web-version is built only once
and then packaged into an `arm64` and an `amd64` nginx-container.
This simplifies the workflow and removes the need to build twice.
Also, the same cache can be reused for both containers.